# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/payout-methods": {
      "get": {
        "operationId": "PayoutMethodController_getPayoutMethods",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PayoutMethod"
        ]
      },
      "post": {
        "operationId": "PayoutMethodController_createPayoutMethod",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePayoutMethodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PayoutMethod"
        ]
      }
    },
...
    "/api/payout-methods/{id}": {
      "get": {
        "operationId": "PayoutMethodController_getPayoutMethod",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PayoutMethod"
        ]
      },
      "put": {
        "operationId": "PayoutMethodController_updatePayoutMethod",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PayoutMethod"
        ]
      },
      "delete": {
        "operationId": "PayoutMethodController_deletePayoutMethod",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PayoutMethod"
        ]
      }
    },
    "/api/payout-methods/{id}/set-default": {
      "post": {
        "operationId": "PayoutMethodController_setDefaultPayoutMethod",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PayoutMethod"
        ]
      }
    },
...
    "/api/payouts/{id}": {
      "get": {
        "operationId": "PayoutController_getPayout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get payout details by ID",
        "tags": [
          "Provider Payouts"
        ]
      },
      "delete": {
        "operationId": "PayoutController_cancelPayout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a pending payout",
        "tags": [
          "Provider Payouts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gripper API",
    "description": "The comprehensive API for the Gripper platform backend.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://api.gripper.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePayoutMethodDto": {
        "type": "object",
        "properties": {}
      },
...
      "PayoutResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "providerId": {
            "type": "string"
          },
          "payoutMethodId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed",
              "cancelled"
            ]
          },
          "gatewayReference": {
            "type": "string"
          },
          "processedAt": {
            "format": "date-time",
            "type": "string"
          },
          "failedAt": {
            "format": "date-time",
            "type": "string"
          },
          "failureReason": {
            "type": "string"
          },
          "payoutMethod": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reference",
          "providerId",
          "amount",
          "currency",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}